+++ test/AccessTest2.java
@@ -18,7 +18,7 @@ public class AccessTest2 {
         Nearest nearest = new Nearest(r, 128);
         RTree rtree = nearest.createRTree();
 
-        int accesos = 0;
+        long accesos = 0;
         for (int i = 0; i < randomListQ.size(); i++) {
             SearchResult result = rtree.search(randomListQ.get(i));
             accesos = accesos + result.getCounter();
@@ -32,7 +32,7 @@ public class AccessTest2 {
         STR str = new STR(r, 128);
         RTree rtree = str.createRTree();
 
-        int accesos = 0;
+        long accesos = 0;
         for (int i = 0; i < randomListQ.size(); i++) {
             SearchResult result = rtree.search(randomListQ.get(i));
             accesos = accesos + result.getCounter();
@@ -58,7 +58,7 @@ public class AccessTest2 {
         HilbertCurve hilbert = new HilbertCurve(r, 128);
         RTree rtree = hilbert.createRTreeH();
 
-        int accesos = 0;
+        long accesos = 0;
         for (int i = 0; i < randomListQ.size(); i++) {
             SearchResult result = rtree.search(randomListQ.get(i));
             accesos = accesos + result.getCounter();
